Rename view
Follow filename conventions
