+++ lib/assets/javascripts/cartodb3/editor/layers/layer-analyses-view.js
