+++ lib/assets/javascripts/cartodb3/editor/layers/layers-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
 var LayerAnalysisViewFactory = require('./layer-analysis-view-factory');
-var LayerAnalysisViews = require('./layer-analysis-views');
+var LayerAnalysesView = require('./layer-analyses-view');
 var AddAnalysisView = require('../../components/modals/add-analysis/add-analysis-view');
 var template = require('./layers.tpl');
 
@@ -88,7 +88,7 @@ module.exports = cdb.core.View.extend({
    * As defined by default-layer-view
    */
   _newAnalysisViews: function (el, layerDefinitionModel) {
-    return new LayerAnalysisViews({
+    return new LayerAnalysesView({
       el: el,
       model: layerDefinitionModel,
       analysisDefinitionNodesCollection: this._analysisDefinitionNodesCollection,
