+++ lib/assets/test/spec/cartodb3/editor/layers/layer-analysis-views.spec.js
@@ -1,10 +1,10 @@
 var AnalysisDefinitionNodesCollection = require('../../../../../javascripts/cartodb3/data/analysis-definition-nodes-collection');
 var LayerDefinitionsCollection = require('../../../../../javascripts/cartodb3/data/layer-definitions-collection');
-var LayerAnalysisViews = require('../../../../../javascripts/cartodb3/editor/layers/layer-analysis-views');
+var LayerAnalysesView = require('../../../../../javascripts/cartodb3/editor/layers/layer-analyses-view');
 var LayerAnalysisViewFactory = require('../../../../../javascripts/cartodb3/editor/layers/layer-analysis-view-factory');
 var createDefaultVis = require('../../create-default-vis');
 
-describe('editor/layers/layer-analysis-views', function () {
+describe('editor/layers/layer-analyses-view', function () {
   beforeEach(function () {
     this.analysisDefinitionNodesCollection = new AnalysisDefinitionNodesCollection(null, {
       sqlAPI: {},
@@ -31,7 +31,7 @@ describe('editor/layers/layer-analysis-views', function () {
           cartocss: 'asd'
         }
       });
-      this.view = new LayerAnalysisViews({
+      this.view = new LayerAnalysesView({
         model: this.layerDefinitionsCollection.get('l1'),
         analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
         modals: {},
@@ -80,7 +80,7 @@ describe('editor/layers/layer-analysis-views', function () {
           name: 'layerB'
         }
       });
-      this.view = new LayerAnalysisViews({
+      this.view = new LayerAnalysesView({
         model: this.layerDefinitionsCollection.get('l1'),
         analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
         layerAnalysisViewFactory: this.layerAnalysisViewFactory
@@ -122,7 +122,7 @@ describe('editor/layers/layer-analysis-views', function () {
             name: 'layerC'
           }
         });
-        this.view = new LayerAnalysisViews({
+        this.view = new LayerAnalysesView({
           model: this.layerDefinitionsCollection.get('l2'),
           analysisDefinitionNodesCollection: this.analysisDefinitionNodesCollection,
           layerAnalysisViewFactory: this.layerAnalysisViewFactory
